Weather api working, login and logout working
The weather api was a hashie has, needing two keys ["key"]["key"] to get the right value. Trolled on hey-9 to get it working.  The registration and login had the same issue, needing to use params[:user][:email] to find by email address.

Code is messy with attempts to use apis, extra gems, and my attempt at creating a profile.